chore: refactor use search into own file
